+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This CHANGELOG file
 - PSR-15 RequestHandlerInterface implementation
 
+### Changed
+
+- Split Server and RequestHandler classes
+
 ### Removed
 
 - handleServerRequest method